Apply allow attribute to a statement, not an expr
This works around the lack of expression attributes by placing it on the
surrounding block instead. This means that the 'nightly' feature is no
longer needed. I left it in place so as not to break users; I'll leave
that up to the owner's judgment whether you want to make a breaking
change or not (or alternatively, emit a warning that the feature is no
longer needed).

Tested with `cargo +stable test` and `cargo +nightly test`.